refactor: renmae process.py to application_process.py
